fixes for elixir Dockerfile
