+++ agario/examples/elixir/README.md
@@ -1,0 +1,37 @@
+# Документация по использованию
+
+Стратегия собирается в escript и запускается с его же помощью. Важно соблюсти следующие требования:
+
+- Одиночный файл
+
+Имя модуля должно быть `Strategy`, имя главной функции -- `main`. Пример стратегии приведен в файле `single-file.ex`.
+
+- Zip-файл
+
+В файле проекта необходимо указать путь, по которому сборка в escript положит его в нужное место. Это место хранится в env-переменной `COMPILED_FILE_PATH`.
+Пример готовой стратегии можно увидеть в папке `strategy`. в файле `mix.exs` есть строчки:
+```
+  ...
+  @output_filename if System.get_env("COMPILED_FILE_PATH"), do: System.get_env("COMPILED_FILE_PATH"), else: "strategy"
+  ...
+  def project do 
+    ...
+    escript: [main_module: Strategy, path: @output_filename],
+    ...
+  end
+```
+Именно эти строки и указывают нужные параметры. Все остальное -- на ваше усмотрение, в том числе имена модулей и проекта.
+
+Нельзя забывать, что в момент сборки на сервере не будет доступен интернет, поэтому все зависимости (hex уже будет стоять) нужно не только указать, но и скачать самостоятельно. То есть, после написания стратегии и указания зависимостей в `mix.exs` вам необходимо выполнить:
+```
+  mix deps.get
+```
+на вашем компьютере. Затем **вся** папка упаковывается в zip-файл и отправляется, вместе со всеми зависимостями. Ограничение на размер, как и в остальных контейнерах, 20 мегабайт.
+
+Чтобы протестировать стратегию локально, запустите
+```
+  MIX_ENV=prod COMPILED_FILE_PATH=strategy mix escript.build
+  escript strategy
+```
+
+Важно, чтобы ни при сборке, ни при запуске ничего не докачивалось из интернета, иначе работать не будет. На сервере запускается с помощью примерно такой же последовательности команд (подробности можно увидеть в соответствующем Dockerfile).
